libmdbx: re-vendor at the v0.14.3 release (0.14.3)
Upstream cut v0.14.3 "Китов" (Kitov) on 2026-08-09, so replace the
engineering-line snapshot vendored earlier in this branch with the tagged
release. Upstream also rewrote its history, so the previously recorded
commit 4ca45169 no longer exists; VERSION.json now records
v0.14.3-0-g251562b2.

On top of the snapshot this brings the fix for the copy-without-compaction
regression that produced a copy without payload, the MinGW static-build
fix for global init and TLS destructors, the cursor-stack rework with the
page stash, spilling/accounting fixes for MDBX_AVOID_MSYNC=ON, and MSVC
C11-atomics workarounds. mdbx.h changes are limited to dropping
MDBX_VERSION_UNSTABLE and adding MDBX_EEXIST -- no signature moved.

Verified on darwin/arm64: go build ./... and go test [-race] ./mdbx
./exp/mdbxpool pass. A throwaway probe confirms a non-compacting
(MDBX_CP_DEFAULTS and MDBX_CP_FORCE_DYNAMIC_SIZE) env copy now round-trips
its payload, where the same probe reads an empty database against the
previously vendored revision.

+++ libmdbx/mdbx_defrag.c
@@ -1,4 +1,4 @@
-/* This file is part of the libmdbx amalgamated source code (v0.14.2-492-g4ca45169 at 2026-07-31T22:58:19+03:00).
+/* This file is part of the libmdbx amalgamated source code (v0.14.3-0-g251562b2 at 2026-08-09T13:18:46+03:00).
  *
  * libmdbx (aka MDBX) is an extremely fast, compact, powerful, embeddedable, transactional key-value storage engine with
  * open-source code. MDBX has a specific set of properties and capabilities, focused on creating unique lightweight
@@ -176,7 +176,7 @@ static int defrag_notify(void *ctx, const MDBX_defrag_result_t *progress) {
     }
 
     if (ctx && tick) {
-      last_report_spenttime = progress->spent_time_dot16;
+      last_report_spenttime = (uint32_t)progress->spent_time_dot16;
       defrag_report_progress(progress, progress_dots++);
     }
     last_progress = *progress;
